Bug 21790: Switch error page template to Bootstrap grid
This patch updates the error page template to use a Bootstrap grid
instead of YUI.

To test, apply the patch and disable Plack and restart Apache if
necessary.

Navigate to a page in Koha which doesn't exist. The error page should
look correct and adapt well to various browser widths.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>