Bug 21805: Duplicate include file in search results template causes JS error
This patch removes a redundant include file, browser-strings.inc, from
the top of the staff client search results template. The
correctly-placed include is in the JavaScript block at the bottom of the
template. The duplication causes a JavaScript error which can be seen in
the browser console.

To test, apply the patch and peform a catalog search in the staff
client. Open the browser console and confirm that there are no errors.

Signed-off-by: Pierre-Marc Thibault <pierre-marc.thibault@inLibro.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>