Bug 21814: System preferences save button can be hidden by language menu
This patch modifies the staff client CSS to give pages enough of a
bottom margin that the language footer will not obscure controls
appearing at the bottom of the page.

To test, apply the patch and regenerate the staff client CSS. You must
have one or more translations installed and enabled.
StaffLangSelectorMode should be set to display the language menu in the
footer.

Go to Administration and search for 'opac.' Scroll to the bottom of the
page and confirm that the last submit button is not hidden by the
language menu.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>
Signed-off-by: Julian Maurice <julian.maurice@biblibre.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>