Bug 21948: Clean up style of item detail page
This patch removes some page-specific style from the item detail page in
the staff client and adds some markup in order to trigger a more
generally-used style.

This patch also adds a font-awesome icon to the "Edit item" link on the
page.

To test, apply the patch and regenerate the staff client CSS. Open a
bibliographic record and view the "Items tab."

Signed-off-by: Hayley Mapley <hayleymapley@catalyst.net.nz>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>