Bug 22035: Improve local cover image browser page
This patch makes changes to the way local cover images are handled on
the image viewer page in the staff client. From this page you can now
delete images or reach the upload images page.

The patch also modifies slightly the process for deleting an image from
the images tab on the bibliographic detail page.

To test, apply the patch and regenerate the staff client CSS.
LocalCoverImages and  AllowMultipleCovers should be enabled.

 - Open the detail view for a record which has multiple cover images
   attached. On the Images tab, click the "Delete image" link for an
   image. After confirming, the link should change to a spinner icon and
   image should fade out.
 - Click one of the cover images to open the image viewer page. The
   thumbnails should each have a delete link which looks like and
   behaves like the delete links on the biblio detail page.
 - Clicking an image thumbnail should load a spinner icon into the area
   where the full-sized image was, which should be replaced by the image
   you chose.
 - There should be a button for uploading images which takes you to the
   correct page for adding images to that record.
 - The page title and breadcrumbs should show the title of the record.

Signed-off-by: Claire Gravely <claire.gravely@bsz-bw.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>