Bug 22195: Change default DataTables configuration to consolidate buttons
This patch changes the default DataTables configuration provided by the
columns settings include file. It consolidates DataTables buttons into
to menus: One for column configuration and one for export options.

This change would not affect tables which aren't set up for column
configuration.

To test, apply the patch and view a page with columns configuration.

 Example: Tools -> Notices & slips
  - Confirm that the buttons appear as expected.
  - Confirm that each button menu performs its task correctly (column
    visiblity, export, print).
  - Test other pages with columns configuration. For instance:
    Acquisitions order search, Currencies administration, Holds queue

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Jose-Mario Monteiro-Santos <jose-mario.monteiro-santos@inLibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>