Bug 22196: Clean up Mana KB administration template
This patch makes language changes to the Mana KB administration
template, and corresponding language changes to the Web Services
preference file.

Some markup errors have been fixed:

 - Correct title and breadcrumbs, copied and pasted from another
   template
 - Add ids to form fields to match their label "for" attributes
 - Make <input>s self-closing
 - Correct class names on alert dialogs
 - Wrap text in JavaScript with _() to make it translatable

Also changed: Some language revisions to the Mana dialogs on the
Administration home page.

To test, apply the patch and go to Administration -> Share content with
Mana KB.

Test the process for enabling and configuring Mana KB. Check that
forms work correctly and that the language of instructions and labels
reads well.

Go to Administration -> System preferences -> Web services, and confirm
that the language changes to the Mana KB preference look good.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>