Bug 21845: Sort of issues in OPAC subscription table
This patch adds a default sort option to the table of subscriptions on
the OPAC detail page. This sorting option matches the sort of data
passed from GetLatestSerials.

To test, apply the patch and locate a subscription in the OPAC which has
multiple issues. They should be sorted by default on publication date.

Signed-off-by: mikael <mikael.Olangcay-Brisebois@inlibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>