Bug 22261: Revise style of DataTables menus
This patch revises our custom DataTables CSS to give a more colorful
style to the columns configuration and export menus. The primary goal of
this change is to make it easier to tell which columns are visible and
which are hidden.

To test, apply the patch and clear your browser cache if necessary.

View a table with columns configuration and export options, e.e. Tools
-> Notices & slip. Test the DataTables menus and confirm that they work
as expected.

Signed-off-by: Liz Rea <wizzyrea@gmail.com>
This is a nice update to the styles.

Signed-off-by: Jose-Mario Monteiro-Santos <jose-mario.monteiro-santos@inLibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>