Bug 21850: (follow-up) Add query information in title attribute
This follow-up adds back the old machine-friendly search details to a
title attribute in case it might be useful for debugging purposes.

To test, apply the patch and perform a search in the OPAC. On the search
results page, hover your mouse over the "Results" link in the
breadcrumbs. The title attribute value should appear with the search
details.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>