Bug 22293: Sticky toolbar making vendor form uneditable
The new floating toolbar plugin doesn't like a particular combination of
markup: A toolbar inside a form which is styled "display:inline." The
staff client CSS styles forms that way globally, so it must be
explicitly overridden in these situations.

Changing the default will have broad consequences, so I propose an
inline style to fix this.

To test, apply the patch and go to Acquisitions. Add or edit a vendor
and confirm that after scrolling down the page the form is still
editable.

Signed-off-by: Andrew Isherwood <andrew.isherwood@ptfs-europe.com>
Signed-off-by: Mikaël Olangcay Brisebois <mikael.olangcay-brisebois@inLibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>