Bug 22351: SCSS conversion broke style on last checked out information
This patch makes a minor change to the staff client SCSS to correct an
error which happened in the SCSS conversion.

To test, apply the patch and regenerate the staff client CSS. Check out
an item to a patron. The box showing information about the item you just
checked out should be flush with the checkout form box.

Signed-off-by: Pierre-Marc Thibault <pierre-marc.thibault@inLibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>