Bug 22432: Star ratings plugin replacement missing from a couple pages
This patch modifies two templates to conform to the change to the start
ratings plugin: Overdrive search results and RecordedBooks search
results.

On the Overdrive page, the star ratings are now correctly displayed
using the new plugin. On the RecordedBooks page, star ratings code has
been removed because it is unused. RecordedBooks doesn't provide ratings
information in their API.

To test you need credentials for both Overdrive and RecordedBooks. Apply
the patch and enable both services.

 - In the OPAC, perform a search that will return results from Overdrive
   and RecordedBooks.
 - Open the search results for each of those services.
   - On the Overdrive search results page, star ratings should appear
     correctly (in read-only mode).
   - On the RecordedBooks page everything should look the same.

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>