Bug 21943: (follow-up) Fix float error with sidebar menu
This patch makes a minor change to the staff client CSS to fix a float
clearing problem when the <main> element contains only a floated element
like <fieldset>. It is a global change, but shouldn't cause unwanted
effects because a non-clearing <main> isn't used in any
float-dependent layouts.

To test, apply the patch and regenerate the staff client CSS. View the
place hold screen and confirm that the sidebar menu looks correct at
narrower browser widths.

Signed-off-by: Liz Rea <wizzyrea@gmail.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>