Bug 22318: (follow-up) Add new options to new table filter form
This patch adds the new language/location options to the form for
filtering entries in the table of news items. Since this markup is
duplicated twice in the template, its contents have been moved to a
BLOCK.

To test, apply the patch and go to Tools -> News. At the top of the
table of news items the "Display location" form should show all options
including OpacNavRight for all your enabled languages.

Confirm that location/language selection still works correctly upon
adding and editing news items.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>