Bug 22525: Navigation arrows on the bottom of Cataloging search pages
This patch adds a second set of pagination links to the cataloging
search result page, at the bottom of the first set of search results.

Also added is a horizontal rule after the second set of links in order
to make it clearer that the links are for the search results above.

To test, apply the patch and go to Cataloging. Perform a search which
will return multiple pages of results. Confirm that the links work
correctly.

Signed-off-by: Marjorie <marjorie.barry-vila@collecto.ca>
Signed-off-by: Bin Wen <bin.wen@inlibro.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>