Bug 14576: (QA follow-up) Tidy up feature documentation
To test:
Apply this patch with all of the others
run the updatedatabase, it should work
install a new Koha, this pref should be installed and look fine in the
interface.
Check the syspref, make sure the text looks ok and makes sense.

Bonus points:
for checking to make sure PROC: _PERM_ appears in the new
pereference if InProcessingToShelvingCart was set
for checking to make sure _ALL_: CART appears in the new preferenec if ReturnToShelvingCart was previously set.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>